+++ node/api_admin_network.go
@@ -24,18 +24,12 @@ package node
 
 import (
 	"context"
-	"encoding/hex"
 	"fmt"
 	"strings"
-	"time"
 
-	"github.com/kaiachain/kaia/common/hexutil"
-	"github.com/kaiachain/kaia/crypto"
-	"github.com/kaiachain/kaia/crypto/bls"
 	"github.com/kaiachain/kaia/networks/p2p"
 	"github.com/kaiachain/kaia/networks/p2p/discover"
 	"github.com/kaiachain/kaia/networks/rpc"
-	"github.com/rcrowley/go-metrics"
 )
 
 // AdminNodeNetworkAPI is the collection of administrative API methods exposed only
@@ -271,214 +265,4 @@ func (api *AdminNodeNetworkAPI) SetMaxSubscriptionPerWSConn(num int32) {
 	logger.Info("Change the max subscription number for a websocket connection",
 		"old", rpc.MaxSubscriptionPerWSConn, "new", num)
 	rpc.MaxSubscriptionPerWSConn = num
-}
-
-// AdminNodeAPI is the collection of administrative API methods exposed over
-// both secure and unsecure RPC channels.
-type AdminNodeAPI struct {
-	node *Node // Node interfaced by this API
-}
-
-// NewAdminNodeAPI creates a new API definition for the public admin methods
-// of the node itself.
-func NewAdminNodeAPI(node *Node) *AdminNodeAPI {
-	return &AdminNodeAPI{node: node}
-}
-
-// Peers retrieves all the information we know about each individual peer at the
-// protocol granularity.
-func (api *AdminNodeAPI) Peers() ([]*p2p.PeerInfo, error) {
-	server := api.node.Server()
-	if server == nil {
-		return nil, ErrNodeStopped
-	}
-	return server.PeersInfo(), nil
-}
-
-// BlsPublicKeyInfoOutput has string fields unlike system.BlsPublicKeyInfo.
-type BlsPublicKeyInfoOutput struct {
-	PublicKey string `json:"publicKey"`
-	Pop       string `json:"pop"`
-}
-
-// NodeInfoOutput extends p2p.NodeInfo with additional fields.
-type NodeInfoOutput struct {
-	p2p.NodeInfo
-
-	// Node address derived from node key
-	NodeAddress string `json:"nodeAddress"`
-
-	// BLS public key information for consensus
-	BlsPublicKeyInfo BlsPublicKeyInfoOutput `json:"blsPublicKeyInfo"`
-}
-
-// NodeInfo retrieves all the information we know about the host node at the
-// protocol granularity.
-func (api *AdminNodeAPI) NodeInfo() (*NodeInfoOutput, error) {
-	server := api.node.Server()
-	if server == nil {
-		return nil, ErrNodeStopped
-	}
-
-	var (
-		nodeKey  = api.node.config.NodeKey()
-		nodeAddr = crypto.PubkeyToAddress(nodeKey.PublicKey)
-
-		blsPriv = api.node.config.BlsNodeKey()
-		blsPub  = blsPriv.PublicKey().Marshal()
-		blsPop  = bls.PopProve(blsPriv).Marshal()
-	)
-
-	info := &NodeInfoOutput{
-		NodeInfo:    *server.NodeInfo(),
-		NodeAddress: nodeAddr.Hex(),
-		BlsPublicKeyInfo: BlsPublicKeyInfoOutput{
-			PublicKey: hex.EncodeToString(blsPub),
-			Pop:       hex.EncodeToString(blsPop),
-		},
-	}
-	return info, nil
-}
-
-// Datadir retrieves the current data directory the node is using.
-func (api *AdminNodeAPI) Datadir() string {
-	return api.node.DataDir()
-}
-
-// DebugNodeAPI is the collection of debugging related API methods exposed over
-// both secure and unsecure RPC channels.
-type DebugNodeAPI struct {
-	node *Node // Node interfaced by this API
-}
-
-// NewDebugNodeAPI creates a new API definition for the public debug methods
-// of the node itself.
-func NewDebugNodeAPI(node *Node) *DebugNodeAPI {
-	return &DebugNodeAPI{node: node}
-}
-
-// Metrics retrieves all the known system metric collected by the node.
-func (api *DebugNodeAPI) Metrics(raw bool) (map[string]interface{}, error) {
-	// Create a rate formatter
-	units := []string{"", "K", "M", "G", "T", "E", "P"}
-	round := func(value float64, prec int) string {
-		unit := 0
-		for value >= 1000 {
-			unit, value, prec = unit+1, value/1000, 2
-		}
-		return fmt.Sprintf(fmt.Sprintf("%%.%df%s", prec, units[unit]), value)
-	}
-	format := func(total float64, rate float64) string {
-		return fmt.Sprintf("%s (%s/s)", round(total, 0), round(rate, 2))
-	}
-	// Iterate over all the metrics, and just dump for now
-	counters := make(map[string]interface{})
-	metrics.DefaultRegistry.Each(func(name string, metric interface{}) {
-		// Create or retrieve the counter hierarchy for this metric
-		root, parts := counters, strings.Split(name, "/")
-		for _, part := range parts[:len(parts)-1] {
-			if _, ok := root[part]; !ok {
-				root[part] = make(map[string]interface{})
-			}
-			root = root[part].(map[string]interface{})
-		}
-		name = parts[len(parts)-1]
-
-		// Fill the counter with the metric details, formatting if requested
-		if raw {
-			switch metric := metric.(type) {
-			case metrics.Counter:
-				root[name] = map[string]interface{}{
-					"Overall": float64(metric.Count()),
-				}
-
-			case metrics.Meter:
-				root[name] = map[string]interface{}{
-					"AvgRate01Min": metric.Rate1(),
-					"AvgRate05Min": metric.Rate5(),
-					"AvgRate15Min": metric.Rate15(),
-					"MeanRate":     metric.RateMean(),
-					"Overall":      float64(metric.Count()),
-				}
-
-			case metrics.Timer:
-				root[name] = map[string]interface{}{
-					"AvgRate01Min": metric.Rate1(),
-					"AvgRate05Min": metric.Rate5(),
-					"AvgRate15Min": metric.Rate15(),
-					"MeanRate":     metric.RateMean(),
-					"Mean":         metric.Mean(),
-					"StdDev":       metric.StdDev(),
-					"Overall":      float64(metric.Count()),
-					"Percentiles": map[string]interface{}{
-						"5":  metric.Percentile(0.05),
-						"20": metric.Percentile(0.2),
-						"50": metric.Percentile(0.5),
-						"80": metric.Percentile(0.8),
-						"95": metric.Percentile(0.95),
-					},
-				}
-
-			default:
-				root[name] = "Unknown metric type"
-			}
-		} else {
-			switch metric := metric.(type) {
-			case metrics.Counter:
-				root[name] = map[string]interface{}{
-					"Overall": float64(metric.Count()),
-				}
-
-			case metrics.Meter:
-				root[name] = map[string]interface{}{
-					"Avg01Min": format(metric.Rate1()*60, metric.Rate1()),
-					"Avg05Min": format(metric.Rate5()*300, metric.Rate5()),
-					"Avg15Min": format(metric.Rate15()*900, metric.Rate15()),
-					"Overall":  format(float64(metric.Count()), metric.RateMean()),
-				}
-
-			case metrics.Timer:
-				root[name] = map[string]interface{}{
-					"Avg01Min": format(metric.Rate1()*60, metric.Rate1()),
-					"Avg05Min": format(metric.Rate5()*300, metric.Rate5()),
-					"Avg15Min": format(metric.Rate15()*900, metric.Rate15()),
-					"Overall":  format(float64(metric.Count()), metric.RateMean()),
-					"Maximum":  time.Duration(metric.Max()).String(),
-					"Minimum":  time.Duration(metric.Min()).String(),
-					"Percentiles": map[string]interface{}{
-						"5":  time.Duration(metric.Percentile(0.05)).String(),
-						"20": time.Duration(metric.Percentile(0.2)).String(),
-						"50": time.Duration(metric.Percentile(0.5)).String(),
-						"80": time.Duration(metric.Percentile(0.8)).String(),
-						"95": time.Duration(metric.Percentile(0.95)).String(),
-					},
-				}
-
-			default:
-				root[name] = "Unknown metric type"
-			}
-		}
-	})
-	return counters, nil
-}
-
-// KaiaNodeAPI offers helper utils
-type KaiaNodeAPI struct {
-	stack *Node
-}
-
-// NewKaiaNodeAPI creates a new Web3Service instance
-func NewKaiaNodeAPI(stack *Node) *KaiaNodeAPI {
-	return &KaiaNodeAPI{stack}
-}
-
-// ClientVersion returns the node name
-func (s *KaiaNodeAPI) ClientVersion() string {
-	return s.stack.Server().Name()
-}
-
-// Sha3 applies the Kaia sha3 implementation on the input.
-// It assumes the input is hex encoded.
-func (s *KaiaNodeAPI) Sha3(input hexutil.Bytes) hexutil.Bytes {
-	return crypto.Keccak256(input)
 }
